requirements.txt
oauthlib>=3.2.2
python-dotenv>=1.0.0
sap-llm-commons[all]>=0.2.0
requests-oauthlib>=1.3.1